Added new DbContext and reorganized Models
