+++ Tulip/Program.cs
@@ -5,18 +5,28 @@ using Tulip.Services.Interfaces;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 
-
 var builder = WebApplication.CreateBuilder(args);
 
 /* Configure Logging */
 builder.Logging.AddConsole();
 
 /* Add Services */
-builder.Services.AddDbContext<ApplicationDbContext>(
-    options => options.UseSqlServer(
-        builder.Configuration.GetConnectionString("DefaultConnection")
-    )
-);
+if (builder.Environment.IsDevelopment())
+{
+    builder.Services.AddDbContext<ApplicationDbContext, DevelopmentDbContext>(
+        options => options.UseSqlServer(
+            builder.Configuration.GetConnectionString("DefaultConnection")
+        )
+    );
+}
+else
+{
+    builder.Services.AddDbContext<ApplicationDbContext>(
+        options => options.UseSqlServer(
+            builder.Configuration.GetConnectionString("DefaultConnection")
+        )
+    );
+}
 
 //builder.Services.AddDatabaseDeveloperPageExceptionFilter();
 
@@ -26,9 +36,16 @@ builder.Services.AddScoped(sp => new HttpClient {
     BaseAddress = new Uri(builder.Configuration.GetValue<string>("APIKey"))
 });
 
-builder.Services
-    .AddIdentity<IdentityUser, IdentityRole>()
-    .AddEntityFrameworkStores<ApplicationDbContext>()
+var identityService = builder.Services.AddIdentity<IdentityUser, IdentityRole>();
+if (builder.Environment.IsDevelopment())
+{
+    identityService.AddEntityFrameworkStores<DevelopmentDbContext>();
+}
+else
+{
+    identityService.AddEntityFrameworkStores<ApplicationDbContext>();
+}
+identityService
     .AddDefaultTokenProviders()
     .AddDefaultUI();
