+++ Tulip/Services/Implementations/TasksService.cs
@@ -2,9 +2,6 @@
 using Tulip.Services.Interfaces;
 using Tulip.Models;
 using Microsoft.EntityFrameworkCore;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Tulip.Services.Implementations
 {
@@ -24,7 +21,7 @@ namespace Tulip.Services.Implementations
 
 		public async Task<TasksResponse> GetTasksResponse(string username)
 		{
-			throw new System.NotImplementedException();
+			throw new NotImplementedException();
 		}
 
 		public async Task<IEnumerable<TasksResponse>> GetResponsePoint(string username)
